Main Arduino Code for ESP32
This code already been test to be used on ESP32-C3 Dual USB with 4MB flash on chip. Any other ESP32, ESP32-S2, ESP32-S3, ESP32-C3 and ESP32-C6 with at least 4MB Flash and WiFi capable should be OK to work. Just be sure that memory partitioning and buffer for the Tensorflow area is suitable for the board.